+++ common/pinot/pinot_client_test.go
@@ -244,93 +244,6 @@ func TestBuildMap(t *testing.T) {
 	}
 }
 
-func TestConvertSearchResultToVisibilityRecord(t *testing.T) {
-	columnName := []string{"WorkflowID", "RunID", "WorkflowType", "DomainID", "StartTime", "ExecutionTime", "CloseTime", "CloseStatus", "HistoryLength", "TaskList", "IsCron", "NumClusters", "UpdateTime", "Attr"}
-	hit := []interface{}{"wfid", "rid", "wftype", "domainid", testEarliestTime, testEarliestTime, testLatestTime, -1, 1, "tsklst", true, 1, testEarliestTime, "{}"}
-	hitWithAttr := []interface{}{"wfid", "rid", "wftype", "domainid", testEarliestTime, testEarliestTime, testLatestTime, 1, 1, "tsklst", true, 1, testEarliestTime, `{"CustomStringField": "customA and customB or customC", "CustomDoubleField": 3.14}`}
-	hitWithAttrMarshalError := []interface{}{"wfid", "rid", "wftype", "domainid", testEarliestTime, testEarliestTime, testLatestTime, 1, 1, "tsklst", true, 1, testEarliestTime, make(chan int)}
-	hitWithAttrUnmarshalError := []interface{}{"wfid", "rid", "wftype", "domainid", testEarliestTime, testEarliestTime, testLatestTime, 1, "1", "tsklst", true, 1, testEarliestTime, `{"CustomStringField": "customA and customB or customC", "CustomDoubleField": 3.14}`}
-
-	closeStatus := types.WorkflowExecutionCloseStatusFailed
-
-	tests := map[string]struct {
-		inputColumnNames         []string
-		inputHit                 []interface{}
-		expectedVisibilityRecord *p.InternalVisibilityWorkflowExecutionInfo
-	}{
-		"Case1: nil result": {
-			inputColumnNames:         nil,
-			inputHit:                 hit,
-			expectedVisibilityRecord: nil,
-		},
-		"Case2-1: marshal system key error case": {
-			inputColumnNames:         columnName,
-			inputHit:                 hitWithAttrMarshalError,
-			expectedVisibilityRecord: nil,
-		},
-		"Case2-2: unmarshal system key error case": {
-			inputColumnNames:         columnName,
-			inputHit:                 hitWithAttrUnmarshalError,
-			expectedVisibilityRecord: nil,
-		},
-		"Case3-1: with everything except for an empty Attr": {
-			inputColumnNames: columnName,
-			inputHit:         hit,
-			expectedVisibilityRecord: &p.InternalVisibilityWorkflowExecutionInfo{
-				DomainID:         "domainid",
-				WorkflowType:     "wftype",
-				WorkflowID:       "wfid",
-				RunID:            "rid",
-				TypeName:         "wftype",
-				StartTime:        time.UnixMilli(testEarliestTime),
-				ExecutionTime:    time.UnixMilli(testEarliestTime),
-				CloseTime:        time.UnixMilli(testLatestTime),
-				Status:           nil,
-				HistoryLength:    1,
-				Memo:             nil,
-				TaskList:         "tsklst",
-				IsCron:           true,
-				NumClusters:      1,
-				UpdateTime:       time.UnixMilli(testEarliestTime),
-				SearchAttributes: map[string]interface{}{},
-				ShardID:          0,
-			},
-		},
-		"Case3-2: with everything": {
-			inputColumnNames: columnName,
-			inputHit:         hitWithAttr,
-			expectedVisibilityRecord: &p.InternalVisibilityWorkflowExecutionInfo{
-				DomainID:         "domainid",
-				WorkflowType:     "wftype",
-				WorkflowID:       "wfid",
-				RunID:            "rid",
-				TypeName:         "wftype",
-				StartTime:        time.UnixMilli(testEarliestTime),
-				ExecutionTime:    time.UnixMilli(testEarliestTime),
-				CloseTime:        time.UnixMilli(testLatestTime),
-				Status:           &closeStatus,
-				HistoryLength:    1,
-				Memo:             nil,
-				TaskList:         "tsklst",
-				IsCron:           true,
-				NumClusters:      1,
-				UpdateTime:       time.UnixMilli(testEarliestTime),
-				SearchAttributes: map[string]interface{}{"CustomStringField": "customA and customB or customC", "CustomDoubleField": 3.14},
-				ShardID:          0,
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.NotPanics(t, func() {
-				visibilityRecord := ConvertSearchResultToVisibilityRecord(test.inputHit, test.inputColumnNames, testlogger.New(t))
-				assert.Equal(t, test.expectedVisibilityRecord, visibilityRecord)
-			})
-		})
-	}
-}
-
 func TestGetInternalListWorkflowExecutionsResponse(t *testing.T) {
 	columnName := []string{"WorkflowID", "RunID", "WorkflowType", "DomainID", "StartTime", "ExecutionTime", "CloseTime", "CloseStatus", "HistoryLength", "Encoding", "TaskList", "IsCron", "NumClusters", "UpdateTime", "Attr"}
 	hit1 := []interface{}{"wfid1", "rid1", "wftype1", "domainid1", testEarliestTime, testEarliestTime, testLatestTime, 1, 1, "encode1", "tsklst1", true, 1, testEarliestTime, "null"}
